+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+### [0.5.0]
+
+### Changed
+- Update to Rider 2022.3 stable
+- Add fix for Fantomas v6.0 alpha
+
 ### [0.5.0-alpha-001]
 
 ### Changed